raft: make election and heartbeat ticks of type int64
In the near future, we want to keep incrementing the electionElapsed
until we hear from a leader. In order to avoid any chances of an integer
overflow, this commit changes the electionElapsed (along with other
symmetrical variables from int to uint64)

Epic: None

Release note: None

+++ pkg/kv/kvserver/replica_proposal.go
@@ -138,11 +138,11 @@ type ProposalData struct {
 
 	// proposedAtTicks is the (logical) time at which this command was
 	// last (re-)proposed.
-	proposedAtTicks int
+	proposedAtTicks int64
 
 	// createdAtTicks is the (logical) time at which this command was
 	// *first* proposed.
-	createdAtTicks int
+	createdAtTicks int64
 
 	// command is the log entry that is encoded into encodedCommand and proposed
 	// to raft. Never mutated.
